Update app so fetching of static content doesn't require authenticated access (so the site-wide alert banner can be displayed before a user logs in)
