+++ app/client/src/components/app.tsx
@@ -1,4 +1,5 @@
 import { useState, useEffect, useCallback } from "react";
+import { render } from "react-dom";
 import {
   BrowserRouter,
   Navigate,
@@ -19,6 +20,7 @@ import "@formio/choices.js/public/assets/styles/choices.min.css";
 // ---
 import { serverBasePath, serverUrl, cloudSpace, fetchData } from "../config";
 import Loading from "components/loading";
+import MarkdownContent from "components/markdownContent";
 import Welcome from "components/welcome";
 import Dashboard from "components/dashboard";
 import ConfirmationDialog from "components/confirmationDialog";
@@ -28,9 +30,84 @@ import NewRebateForm from "routes/newRebateForm";
 import ExistingRebateForm from "routes/existingRebateForm";
 import NotFound from "routes/notFound";
 import { useUserState, useUserDispatch } from "contexts/user";
+import { useContentState, useContentDispatch } from "contexts/content";
 import { useDialogDispatch, useDialogState } from "contexts/dialog";
 
-type HelpdeskAccess = "idle" | "pending" | "success" | "failure";
+// Custom hook to fetch static content
+function useFetchedContent() {
+  const dispatch = useContentDispatch();
+
+  useEffect(() => {
+    dispatch({ type: "FETCH_CONTENT_REQUEST" });
+    fetchData(`${serverUrl}/api/content`)
+      .then((res) => {
+        const {
+          siteAlert,
+          helpdeskIntro,
+          allRebateFormsIntro,
+          allRebateFormsOutro,
+          newRebateFormIntro,
+          newRebateFormDialog,
+          existingDraftRebateFormIntro,
+          existingSubmittedRebateFormIntro,
+        } = res;
+        dispatch({
+          type: "FETCH_CONTENT_SUCCESS",
+          payload: {
+            siteAlert,
+            helpdeskIntro,
+            allRebateFormsIntro,
+            allRebateFormsOutro,
+            newRebateFormIntro,
+            newRebateFormDialog,
+            existingDraftRebateFormIntro,
+            existingSubmittedRebateFormIntro,
+          },
+        });
+      })
+      .catch((err) => {
+        dispatch({ type: "FETCH_CONTENT_FAILURE" });
+      });
+  }, [dispatch]);
+}
+
+// Custom hook to display a site-wide alert banner
+function useSiteAlertBanner() {
+  const { content } = useContentState();
+
+  useEffect(() => {
+    if (content.status !== "success") return;
+    if (content.data?.siteAlert === "") return;
+
+    const siteAlert = document.querySelector(".usa-site-alert");
+    if (!siteAlert) return;
+
+    siteAlert.setAttribute("aria-label", "Site alert");
+    siteAlert.classList.add("usa-site-alert--emergency");
+
+    render(
+      <div className="usa-alert">
+        <MarkdownContent
+          className="usa-alert__body"
+          children={content.data?.siteAlert || ""}
+          components={{
+            h1: (props) => (
+              <h3 className="usa-alert__heading">{props.children}</h3>
+            ),
+            h2: (props) => (
+              <h3 className="usa-alert__heading">{props.children}</h3>
+            ),
+            h3: (props) => (
+              <h3 className="usa-alert__heading">{props.children}</h3>
+            ),
+            p: (props) => <p className="usa-alert__text">{props.children}</p>,
+          }}
+        />
+      </div>,
+      siteAlert
+    );
+  }, [content]);
+}
 
 // Custom hook to display the CSB disclaimer banner for development/staging
 function useDisclaimerBanner() {
@@ -53,20 +130,6 @@ function useDisclaimerBanner() {
 
     siteAlert.insertAdjacentElement("beforebegin", banner);
   }, []);
-}
-
-// Custom hook to check if user should have access to helpdesk pages
-export function useHelpdeskAccess() {
-  const [helpdeskAccess, setHelpdeskAccess] = useState<HelpdeskAccess>("idle");
-
-  useEffect(() => {
-    setHelpdeskAccess("pending");
-    fetchData(`${serverUrl}/api/helpdesk-access`)
-      .then((res) => setHelpdeskAccess("success"))
-      .catch((err) => setHelpdeskAccess("failure"));
-  }, []);
-
-  return helpdeskAccess;
 }
 
 // Custom hook to set up inactivity timer to auto-logout user if they're inactive for >15 minutes
@@ -153,6 +216,22 @@ function useInactivityDialog(callback: () => void) {
   }, [dialogShown, heading, logoutTimer, dispatch]);
 }
 
+// Custom hook to check if user should have access to helpdesk pages
+export function useHelpdeskAccess() {
+  const [helpdeskAccess, setHelpdeskAccess] = useState<
+    "idle" | "pending" | "success" | "failure"
+  >("idle");
+
+  useEffect(() => {
+    setHelpdeskAccess("pending");
+    fetchData(`${serverUrl}/api/helpdesk-access`)
+      .then((res) => setHelpdeskAccess("success"))
+      .catch((err) => setHelpdeskAccess("failure"));
+  }, []);
+
+  return helpdeskAccess;
+}
+
 // Wrapper Component for any routes that need authenticated access
 function ProtectedRoute({ children }: { children: JSX.Element }) {
   const { pathname } = useLocation();
@@ -201,6 +280,8 @@ function ProtectedRoute({ children }: { children: JSX.Element }) {
 }
 
 export default function App() {
+  useFetchedContent();
+  useSiteAlertBanner();
   useDisclaimerBanner();
 
   return (
